Fix for #23: Adding extremely simple "hello world" MVP to demonstrate svelte-retag in most basic form. Relocating Vercel site files to subdirectory under ./demo/vercel so this can live under ./demo/hello-world.
