+++ demo/hello-world/.vscode/extensions.json
