+++ demo/hello-world/README.md
@@ -1,0 +1,15 @@
+# svelte-retag: Hello World
+
+This is an **MVP** demo that shows `svelte-retag` in action in its simplest form in local development with a
+`HelloWorld.svelte` component implemented as a `<hello-world>` custom element.
+
+To try it out yourself, make sure you have [Git](https://git-scm.com/downloads) and [Node](https://nodejs.org/) installed and then run:
+
+```bash
+git clone https://github.com/patricknelson/svelte-retag.git
+cd svelte-retag/demo/hello-world
+npm install
+npm run dev
+```
+
+Then open your browser to [http://localhost:5173/](http://localhost:5173/) to view the demo.
