+++ demo/hello-world/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/svelte-retag-favicon.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>svelte-retag: Hello World</title>
+  </head>
+  <body>
+    <hello-world greetperson="👩‍🚀"></hello-world>
+    <script type="module" src="/src/main.js"></script>
+  </body>
+</html>
