+++ demo/hello-world/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "svelte-retag-hello-world",
+  "private": true,
+  "version": "0.0.1",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "svelte": "^4.2.3",
+    "vite": "^5.0.0"
+  },
+  "dependencies": {
+    "svelte-retag": "^1.3.2"
+  }
+}
