+++ demo/hello-world/src/HelloWorld.svelte
@@ -1,0 +1,5 @@
+<script>
+	export let greetPerson = 'World';
+</script>
+
+<h1>Hello {greetPerson}!</h1>
