+++ demo/hello-world/src/main.js
@@ -1,0 +1,7 @@
+import svelteRetag from 'svelte-retag';
+import HelloWorld from './HelloWorld.svelte';
+
+svelteRetag({
+	component: HelloWorld,
+	tagname: 'hello-world',
+});
