+++ demo/vercel/hydratable.html
@@ -51,7 +51,7 @@
 				<!-- Features -->
 				<tab-panel data-svelte-retag-hydratable=""><svelte-retag><div class="svelte-1xkiur0 active"><svelte-retag-default>
 					<features-info data-svelte-retag-hydratable=""><svelte-retag><p>View the
-						<a href="https://github.com/patricknelson/svelte-retag/blob/main/demo/index.html" target="_blank" rel="noopener">page
+						<a href="https://github.com/patricknelson/svelte-retag/blob/main/demo/vercel/index.html" target="_blank" rel="noopener">page
 							source 🚀</a>. This demo is implemented entirely using via <code>svelte-retag</code> with all Svelte components
 						organized on the page using custom elements (or “web components”). Featuring:</p> <ul><li><strong>Slots</strong> (default and named slots)</li> <li><strong>Nesting</strong> within slots</li> <li><strong>Context</strong> (including nested context, see “Examples” tab)</li> <li><strong>Vite HMR updates</strong> (if launched locally)</li> <li>SSR/SSG via <strong>hydration</strong> <em>(experimental)</em></li></ul> <h3>Performance</h3> <p>This demo also shows the differences between deferred Modules and IIFE, particularly with regard to CLS (Cumulative
 						Layout Shift). Use the <strong>Demos</strong> navigation above and compare each one by opening DevTools, disabling
