+++ demo/vercel/src/lib/tabs/TabsWrapper.svelte
