+++ demo/vercel/src/main.js
@@ -1,4 +1,4 @@
-import svelteRetag from '../../index.js';
+import svelteRetag from '../../../index.js';
 
 import './app.css';
 import App from './App.svelte';
